+++ isExtractableFile.test.mjs
@@ -1,9 +1,15 @@
+// @ts-check
+
 import { strictEqual } from "assert";
 import revertableGlobals from "revertable-globals";
 import ReactNativeFile from "./ReactNativeFile.mjs";
 import isExtractableFile from "./isExtractableFile.mjs";
 import assertBundleSize from "./test/assertBundleSize.mjs";
 
+/**
+ * Adds `isExtractableFile` tests.
+ * @param {import("test-director").default} tests Test director.
+ */
 export default (tests) => {
   tests.add("`isExtractableFile` bundle size.", async () => {
     await assertBundleSize(